code/requirements.txt
jupyter
matplotlib
numpy
notebook
opencv-python
pandas
pillow
pipreqs
scikit-learn
scip
tensorboard
torc
torchaudio
torchdata
torchtext
torchvision
tqdm

